DataStream::skip should take unsigned value, size_t -> long to preserve previous fix (but can't use ssize_t)

+++ OgreMain/include/OgreZip.h
@@ -111,7 +111,7 @@ namespace Ogre {
         /// @copydoc DataStream::skipLine
         size_t skipLine(const String& delim = "\n");
         /// @copydoc DataStream::skip
-        void skip(size_t count);
+        void skip(long count);
         /// @copydoc DataStream::seek
         void seek( size_t pos );
         /// @copydoc DataStream::eof
